refactoring

+++ src/TradingSystem.Binance/Market/BinanceFuturesMarketClient.cs
@@ -12,7 +12,7 @@ public sealed class BinanceFuturesMarketClient(HttpClient httpClient) : IBinance
         using var response  =  await httpClient.GetAsync($"/fapi/v1/premiumIndex?symbol = {Uri.EscapeDataString(symbol.ToUpperInvariant())}",  cancellationToken);
         var content  =  await response.Content.ReadAsStringAsync(cancellationToken);
         if (!response.IsSuccessStatusCode)
-            throw new BinanceApiException((int)response.StatusCode,  content);
+            throw new BinanceApiException(response.StatusCode,  content);
         using var document  =  JsonDocument.Parse(content);
         var value  =  document.RootElement.GetProperty("markPrice").GetString();
         return decimal.TryParse(value,  NumberStyles.Any,  CultureInfo.InvariantCulture,  out var price)  &&  price > 0
